refactor: 불필요기능(오늘사용) 제거, 로직 훅으로 분리
